+++ __tests__/label-widget.js
@@ -1,4 +1,4 @@
-import { Component, options } from '../src/shamUI';
+import { Component } from '../src/shamUI';
 import { renderComponent, expectRenderedText, onRenderComplete } from './helpers';
 
 afterEach( () => {
@@ -6,8 +6,6 @@ afterEach( () => {
 } );
 
 class Label extends Component {
-    @options types = [ 'label' ];
-
     render() {
         this.container.innerHTML = this.options.text();
     }
@@ -34,34 +32,6 @@ it( 'render (ONLY_IDS)', async() => {
     } );
     await onRenderComplete( UI => UI.render.ONLY_IDS( 'dummy' ) );
     expectRenderedText( 'dummy (second render)' );
-} );
-
-it( 'first render (ONLY_TYPES)', async() => {
-    expect.assertions( 2 );
-    const mock = jest
-        .fn()
-        .mockReturnValueOnce( 'dummy (first render)' )
-        .mockReturnValueOnce( 'dummy (second render)' );
-    await renderComponent( Label, {
-        text: mock
-    } );
-    expectRenderedText( 'dummy (first render)' );
-    await onRenderComplete( UI => UI.render.ONLY_TYPES( 'label' ) );
-    expectRenderedText( 'dummy (second render)' );
-} );
-
-it( 'render (not exists type)', async() => {
-    expect.assertions( 2 );
-    const mock = jest
-        .fn()
-        .mockReturnValueOnce( 'dummy (first render)' )
-        .mockReturnValueOnce( 'dummy (second render)' );
-    await renderComponent( Label, {
-        text: mock
-    } );
-    expectRenderedText( 'dummy (first render)' );
-    await onRenderComplete( UI => UI.render.ONLY_TYPES( 'non-exits' ) );
-    expectRenderedText( 'dummy (first render)' );
 } );
 
 it( 'render (not exists id)', async() => {